dante algorithms are now latentscorers, subclass of posthocUQ with optional latent feature extraction
